Выносит константы в отдельный модуль
